ImmutabeOrderedSet: improvements in space&time, some fixes, some tests
-the find method was redefined to return the index that the search
should continue with instead of the one it ended on.
the contains method now uses the index returned by the find method to
check what value is in the keys array and determine if the search was
successful or not.
this change makes the code slightly simpler and avoids the use of an
instance variable to pass information between methods.
-an assertion was added in the process of the change that detected an
error in the performace testing method
-the performance testing was fixed. the data was not passed presorted,
which is incorrect.
-the performance testing was also improved in that it now tests random
access pattern, which are the worst case scenario in terms of temporal
locality of reference for both the binary search algorithm and the
vanembe Boas tree layout. this reveals the spacial locality performance
of the two algorithms:
-for random access pattern on random data, the set now performs
significantly faster than a binary search (Arrays.binarySearch) on the
keys even for input sizes that still fit into main memory. on an intel
core i5 3330S 6M cache processor for 2^25 keys, the set outperformed
binary search by a factor of 2.7
-in the best case for binary search, the immutable set performs only
very slightly faster. (ca 2-10%)
-this demonstrates the superior cache usage predicted by the cache
oblivious analysis.